Refactor project

+++ src/main/kotlin/data/remote/dataSource/WeatherDataSourceImp.kt
@@ -11,7 +11,7 @@ class WeatherDataSourceImp(
     private val client: HttpClient, private val json: Json, private val baseUrl: String
 ) : IWeatherDataSource {
 
-    override suspend fun getWeatherByLocation(latitude: Float, longitude: Float): WeatherResponse {
+    override suspend fun getWeatherByLocation(latitude: Double, longitude: Double): WeatherResponse {
         val response = safeApiCall {
             client.get(
                 "$baseUrl?latitude=$latitude&longitude=$longitude&hourly=temperature_2m," + "wind_speed_10m,showers,cloud_cover&forecast_days=1"
